test
WE ARE NERDSMITH, and so are you! Welcome to our public server, friend. Thank you for joining us!

This is a quick orientation to our community features. You will be able to customize your Nerdsmith experience and choose what topics and groups you see. No more cluttered Discord servers full of dozens of notifications or muted channels!

This shouldn't take more than a minute or two. :smiley:  Just a few steps to begin your Nerdsmith journey! Look for a new channel appearing to the LEFT on the channel list at each step.

When you are ready to begin your journey with us, type !start
==============================
(Any commands referenced on these pages are not yet active, so typing them does nothing. But only for now :smiley: )